Replace old variable name in helpers.lua to fix "Attempt to index a nil value" crash during breeding method
While breeding there's a constantly check for the current egg state of each party member to see if it's time to release all lvl 1 pokemon that weren't the target.

A variable "new_eggs" from a previous version of helpers.lua was renamed, but not everywhere. So it was being used in an ipairs call without being initialized. Simply replacing the new_eggs with party_egg_states fixes the "Attempt to index a nil value" crash.